# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Parcial 2 - Elecciones 2023'","description":"'Parcial 2 - Elecciones 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1'"},"servers":[{"url":"http://localhost:8080","description":"'Parcial 2 - Elecciones 2023'"}],"paths":{"/secciones":{"get":{"tags":["seccion-controller"],"operationId":"getSecciones","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeccionDTO"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/distritos":{"get":{"tags":["distrito-controller"],"operationId":"getDistritos","parameters":[{"name":"distrito_nombre","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DistritoDTO"}}}}}}}},"/cargos":{"get":{"tags":["cargo-controller"],"operationId":"getCargoDistrito","parameters":[{"name":"distrito_id","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CargoDistrito"}}}}}}}},"components":{"schemas":{"SeccionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"DistritoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"CargoDistrito":{"type":"object","properties":{"distrito":{"$ref":"#/components/schemas/Distrito"},"cargos":{"type":"array","items":{"$ref":"#/components/schemas/Cargos"}}}},"Cargos":{"type":"object","properties":{"cargoId":{"type":"integer","format":"int32"},"cargoNombre":{"type":"string"},"distritoId":{"type":"integer","format":"int32"}}},"Distrito":{"type":"object","properties":{"distritoId":{"type":"integer","format":"int64"},"distritoNombre":{"type":"string"}}}}}}