# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/SeccionDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class SeccionDTO {
    private Long id;
    private String nombre;
}